requirements.txt
aiogram==3.17
sqlalchemy==2.0.23
alembic==1.13.1
pyyaml==6.0.1
aiofiles==23.2.1